app structure and room entities creation
