+++ app/src/main/java/com/a1ishka/bankaccount/data/Account.kt
@@ -1,13 +1,14 @@
 package com.a1ishka.bankaccount.data
 
 data class Account(
+    val accountId: Long,
     val name: String,
     val accountNumber: Long,
     val cardNumber: Long
 )
 
 val accountData = listOf(
-    Account("My first account", 91212192291221, 5558794679461234),
-    Account("For travelling", 91212192291221, 5558794679461234),
-    Account("Saving Account", 91212192291221, 5558794679461234)
+    Account(3, "My first account", 91212192291221, 5558794679461234),
+    Account(4, "For travelling", 91212192291221, 5558794679461234),
+    Account(5, "Saving Account", 91212192291221, 5558794679461234)
 )
