+++ app/src/main/java/com/a1ishka/bankaccount/data/entity/AccountEntity.kt
@@ -1,0 +1,13 @@
+package com.a1ishka.bankaccount.data.entity
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity
+data class AccountEntity(
+    @PrimaryKey(autoGenerate = true)
+    val accountId: Long? = 0,
+    val name: String,
+    val accountNumber: Long,
+    val cardNumber: Long
+)
