+++ app/src/main/java/com/a1ishka/bankaccount/presentation/transaction/TransactionEvent.kt
@@ -1,0 +1,17 @@
+package com.a1ishka.bankaccount.presentation.transaction
+
+import com.a1ishka.bankaccount.data.entity.TransactionEntity
+
+sealed interface TransactionEvent {
+    object SaveTransaction : TransactionEvent
+    data class FilterTransactions(val accountId: Long, val startDate: String, val endDate: String) :
+        TransactionEvent
+
+    data class DeleteTransaction(val transaction: TransactionEntity) : TransactionEvent
+    data class SetApplier(val applier: String) : TransactionEvent
+    data class SetNumber(val number: String) : TransactionEvent
+    data class SetDate(val date: String) : TransactionEvent
+    data class SetStatus(val status: String) : TransactionEvent
+    data class SetAmount(val amount: String) : TransactionEvent
+    data class SetAccount(val accountId: Long) : TransactionEvent
+}
